+++ CHANGELOG.md
@@ -7,11 +7,12 @@ This file contains the list of changes made to pyjoulescope_ui.
 
 ## 1.2.6
 
-2025 Feb 11 [in progress]
+2025 Feb 12 [in progress]
 
 * Fixed JLS annotations not displaying from primary JLS file  #299
 * Fixed Trigger widget markers placed slightly early  #300
 * Upgraded Windows folder discovery from CSIDL to Known Folder.
+* Added precision to Multimeter widget and Value widget  #301 
 
 
 ## 1.2.5